+++ geotrek/api/tests/test_v2.py
@@ -6740,7 +6740,7 @@ class SensitivityAPIv2Test(TrekkingManagerTest):
                     "url": "http://url.com",
                 },
             ],
-            "provider": "",
+            "provider": None,
             "structure": self.structure.pk,
             "published": True,
         }