+++ geotrek/common/models.py
@@ -22,6 +22,7 @@ from paperclip.validators import FileMimetypeValidator
 from PIL import Image
 
 from geotrek.authent.models import StructureOrNoneRelated
+from geotrek.common.utils.validators import validate_html_template
 
 from .managers import AccessibilityAttachmentManager
 from .mixins.models import OptionalPictogramMixin, PictogramMixin, TimeStampedModelMixin
@@ -476,3 +477,25 @@ class AccessMean(TimeStampedModelMixin):
 
     def __str__(self):
         return self.label
+
+
+class Provider(TimeStampedModelMixin):
+    name = models.CharField(verbose_name=_("Name"), max_length=1024, unique=True)
+    link_template = models.TextField(
+        verbose_name=_("Link template"),
+        blank=True,
+        default="",
+        help_text=_("HTML template for external source links"),
+        validators=[validate_html_template],
+    )
+    copyright = models.CharField(
+        max_length=1024, blank=True, default="", verbose_name="Copyright"
+    )
+
+    class Meta:
+        verbose_name = _("Provider")
+        verbose_name_plural = _("Providers")
+        ordering = ("name",)
+
+    def __str__(self):
+        return self.name