+++ geotrek/sensitivity/models.py
@@ -13,6 +13,7 @@ from pyopenair.factory import wkt2openair
 from geotrek.authent.models import StructureRelated
 from geotrek.common.mixins.models import (
     AddPropertyMixin,
+    ExternalSourceMixin,
     GeotrekMapEntityMixin,
     NoDeleteMixin,
     OptionalPictogramMixin,
@@ -119,6 +120,7 @@ class SensitiveArea(
     TimeStampedModelMixin,
     NoDeleteMixin,
     AddPropertyMixin,
+    ExternalSourceMixin,
 ):
     geom = models.GeometryField(srid=settings.SRID)
     geom_buffered = models.GeometryField(srid=settings.SRID, editable=False)
@@ -135,12 +137,6 @@ class SensitiveArea(
     )
     description = models.TextField(verbose_name=_("Description"), blank=True)
     contact = models.TextField(verbose_name=_("Contact"), blank=True)
-    eid = models.CharField(
-        verbose_name=_("External id"), max_length=1024, blank=True, null=True
-    )
-    provider = models.CharField(
-        verbose_name=_("Provider"), db_index=True, max_length=1024, blank=True
-    )
     rules = models.ManyToManyField(Rule, verbose_name=_("Rules"), blank=True)
 
     objects = SensitiveAreaManager()