+++ geotrek/signage/templates/signage/sql/post_90_defaults.sql
@@ -29,7 +29,6 @@ ALTER TABLE signage_signage ALTER COLUMN description SET DEFAULT '';
 ALTER TABLE signage_signage ALTER COLUMN published SET DEFAULT FALSE;
 -- publication_date
 -- structure
-ALTER TABLE signage_signage ALTER COLUMN provider SET DEFAULT '';