+++ docs/changelog.rst
@@ -5,9 +5,14 @@ CHANGELOG
 2.116.0+dev    (XXXX-XX-XX)
 ----------------------------
 
+**Improvements**
+
+- Change provider Charfield for a foreign key
+
 **Bug fixes**
 
 - Fix parser crash after exception during attachment download (refs #4807)
+
 
 
 2.116.0    (2025-07-03)